extract claims from jwt token
