+++ pkg/handlers/post.go
@@ -7,7 +7,6 @@ import (
 	"log"
 	"net/http"
 	"net/url"
-	"regexp"
 	"strings"
 )
 
@@ -16,7 +15,7 @@ func (authHandler *AuthHandler) SignIn(responseWriter http.ResponseWriter, reque
 	urlPath := "http://localhost:8080/auth/realms/ubivius/protocol/openid-connect/token"
 
 	requestbody, _ := ioutil.ReadAll(request.Body)
-	log.Println(string(requestbody))
+
 	username := extractValue(string(requestbody), "username")
 	password := extractValue(string(requestbody), "password")
 
@@ -40,19 +39,25 @@ func (authHandler *AuthHandler) SignIn(responseWriter http.ResponseWriter, reque
 	}
 	defer resp.Body.Close()
 
-	log.Println("response Status:", resp.Status)
 	body, _ := ioutil.ReadAll(resp.Body)
 	access_token := extractValue(string(body), "access_token")
-	log.Println("access_token: ", access_token)
 
-	player := map[string]string{"username": username, "access_token": access_token}
+	claims := extractClaims(access_token)
+	// do something with decoded claims
+	/*for key, val := range claims {
+		log.Print(val)
+		log.Print(key)
+	}*/
+	log.Print(claims["sub"])
+	userId := claims["sub"]
+
+	player := map[string]string{"userId": string(userId), "access_token": access_token}
 	playerJson, _ := json.Marshal(player)
 
 	err = json.NewEncoder(responseWriter).Encode(playerJson)
 	if err != nil {
 		http.Error(responseWriter, "Unable to complete SignIn", http.StatusInternalServerError)
 	}
-
 }
 
 func (authHandler *AuthHandler) SignUp(responseWriter http.ResponseWriter, request *http.Request) {
@@ -96,42 +101,4 @@ func (authHandler *AuthHandler) SignUp(responseWriter http.ResponseWriter, reque
 	if err != nil {
 		http.Error(responseWriter, "Unable to complete SignIn", http.StatusInternalServerError)
 	}
-}
-
-func GetAdminAccessToken() string {
-	urlPath := "http://localhost:8080/auth/realms/ubivius/protocol/openid-connect/token"
-
-	data := url.Values{}
-	data.Set("client_id", "ubivius-client")
-	data.Set("grant_type", "client_credentials")
-	data.Set("client_secret", "7d109d2b-524f-4351-bfda-44ecad030eef")
-
-	req, err := http.NewRequest("POST", urlPath, strings.NewReader(data.Encode()))
-	if err != nil {
-		panic(err)
-	}
-	req.Header.Add("Content-Type", "application/x-www-form-urlencoded")
-	client := &http.Client{}
-	resp, err := client.Do(req)
-	if err != nil {
-		panic(err)
-	}
-	defer resp.Body.Close()
-
-	log.Println("Admintoken response Status:", resp.Status)
-	body, _ := ioutil.ReadAll(resp.Body)
-	admin_token := extractValue(string(body), "access_token")
-	return admin_token
-}
-
-// extracts the value for a key from a JSON-formatted string
-// body - the JSON-response as a string. Usually retrieved via the request body
-// key - the key for which the value should be extracted
-// returns - the value for the given key
-func extractValue(body string, key string) string {
-	keystr := "\"" + key + "\":[^,;\\]}]*"
-	r, _ := regexp.Compile(keystr)
-	match := r.FindString(body)
-	keyValMatch := strings.Split(match, ":")
-	return strings.ReplaceAll(keyValMatch[1], "\"", "")
 }
